diff/push: Add new options for conditional behavior based on table size
The new --allow-below-size option makes Skeema more permissive for altering
tables under the specified size. These tables will be treated as if
both --allow-drop-table and --allow-drop-column are enabled. This allows for
rapid development when altering a new table before it's in use, or dropping a
table that was never in use.

The new --alter-wrapper-min-size option causes --alter-wrapper to be ignored
for tables under the specified size. This permits skipping the overhead of
external OSC tools when modifying small tables.

The new {SIZE} placeholder variable is replaced with the size of the table, as
determined by querying information_schema.

When --debug is in use, diff/push add new logging relating to the conditional
checks of the above variables.