Begin laying groundwork for service discovery and sharding
This commit adds Dir.Instances(), which returns can potentially return multiple
hosts for a dir, depending on its configuration: If the hosts option value is
wrapped in backticks, it will be interpreted as a shellout command-line, and
the result will be split on delimiters to obtain multiple hostnames.

This behavior is not actively in-use or documented yet though. All callsites
currently still go through Dir.FirstInstance(), which only ever returns one
host. Subsequent commits will make further changes to permit full use of
multiple hosts, with concurrent execution.