Allow host option to be a comma-separated list of hosts
In addition to supporting a single static host and a shellout of multiple hosts,
allow the host option to be set to a static list of hosts.

This will be documented in a future commit, once the schema option follows suit.

This commit also fixes a bug in port parsing. Previously, if the host option
returned multiple hosts, instances without explicit ports specified were
inheriting the port of the previous instance in some situations.

The github.com/skeema/mycli dep was updated to obtain Config.GetSlice support.