diff/push: Improve error-handling in multi-instance scenarios
If there's a fatal problem found when obtaining the schema from the filesystem,
don't bother generating multiple Targets that all have the same error. Instead
track it as a single errored Target, with no Instance, so that it gets displayed
as a problem with the entire directory.

In the process, the logic formerly in Target.obtainSchemaFromDir() was moved to
new method Dir.TargetTemplate().