Support dirs that map to multiple schema names for identical schema
This commit adds support for sharded topologies that put multiple database
schemas on the same instance, where all schemas have the same set of tables.

The "schema" option may now be set to any of the following:
* single static schema name: works as before
* multiple schema names: comma-separated list
* all non-system schemas on an instance: literal value "*"
* parsed shellout: backtick-wrapped command to execute

These will be documented in a subsequent commit. `skeema pull` still needs an
additional improvement to handle this setup properly in its detect-new-schemas
logic.

This commit also updates the github.com/skeema/tengo dep, to get the new
Instance.SchemasByName() method, and also to add MySQL 5.7+'s sys schema to the
list of ignored system schemas.