pull: improvements to handle dirs that map to multiple schemas
When detecting new schemas on an instance, properly handle multi-schema dirs.
However, abort detection if a subdir has schema "*", since this means all
schemas on the instance are identical and already map to that dir.

Improve output of pull to indicate which instance and schema is being used
for each dir. This helps clarify in the multi-schema and/or multi-instance
case. (pull always only uses the first instance and first schema for each dir.)

Fix a bug in pull (and init) where an error message was not being displayed
properly due to nil pointer dereference.