Table.Diff(): Support changing default charset and/or collation
These changes now generate the appropriate ALTER TABLE clause. (Previously,
this operation was unimplemented and a panic would intentionally be triggered.)

This required some changes to the semantics of Column.CharacterSet and the
Column.Definition() method. CharacterSet is now always populated for textual
column types. Definition() now optionally takes a Table arg, and omits the
CHARACTER SET clause when the supplied Table has the same collation as the
Column. This allows us to match the behavior of SHOW CREATE TABLE but still
retain knowledge of the column's character set, even if it matches the table's,
since we need to know the actual value in case the table's default changes!