Support schema default charset/collation, and schema DDL statements in diffs
This commit adds new field SchemaDiff.SchemaDDL, which may contain a single
statement for creating, dropping, or altering the charset/collation of a schema.

Instance.CreateSchema() also now has the ability to specify default charset and/
or collation for the new schema. This required a breaking API change (2 new
params for the method).

Schema.CreateStatement() now includes clauses for default character set and
collation when appropriate.