New methods for interacting with schema default charset and collation
Schema.AlterStatement() generates an ALTER DATABASE query based on the supplied
new charset and collation.

Schema.OverridesServerCharSet() provides a way to check if a schema is using the
same charset and/or collation as the server default.

Instance.AlterSchema() changes the default charset and/or collation of a schema
on an instance.