Replace ColumnDefaultCurrentTimestamp with ColumnDefaultExpression()
This is necessary for supporting fractional second precision in temporal types,
which is available in MySQL/Percona 5.6+ and MariaDB 5.3+.

MariaDB 10.2's general-purpose default expressions are not yet supported, but
this change will eventually be compatible with them as well.

This commit also updates docs to explicitly mention MariaDB is now a supported
database for Go La Tengo, with some MariaDB-specific features not yet supported
in table diffs.