diff/push: improve debug logging for unsupported table-diff edge case
With --debug enabled, when diff/push encounters a scenario in which Go La Tengo
cannot generate a diff for a table, Skeema ordinarily already displays a textual
diff of the actual vs expected CREATE TABLE statements. However, it did not
previously have a notion of unsupported *table-diffs* -- e.g., cases where Go La
Tengo can fully parse both tables and generate CREATE TABLE statements for them,
but cannot generate a corresponding ALTER TABLE for them. This situation is now
handled as well.

This is really just a safety/assertion improvement, as this scenario is not
expected to actually come up.