Add StatementModifiers.AllowUnsafe field, unifying old AllowDrop fields
This commit replaces StatementModifiers.AllowDropTable and
StatementModifiers.AllowDropColumn with a new unified
StatementModifiers.AllowUnsafe. Each type of TableAlterClause now has an
Unsafe() method to indicate whether or not the particular ALTER clause involves
a potentially destructive operation. For most structs implementing
TableAlterClause, the result of Unsafe() is a constant true or false, though for
ModifyColumn it depends on the specific modification being made.

This commit also changes a few field names in other structs for consistency
purposes.