Simplify optional service discovery configuration
Previously, Skeema interpreted backtick-wrapped host values as external commands
to execute. This behavior was difficult to configure, as it relied on using
directory names as lookup keys in the service discovery command-line.

This commit moves that logic to a new option, host-wrapper. With that option
configured, the host option can now be used as a lookup key to be supplied
command-line. This permits host-wrapper to be set in a single global location,
and then host can still be set on individual directories as desired. This also
permits `skeema init` and `skeema add-environment` to pass their supplied --host
values to the configured service discovery client as well.

This commit also removes old command-line placeholder variables HOSTDIR,
SCHEMADIR, and DIRPARENT as they are no longer needed and were overly complex.