Ignore views in schemas without breaking
The addition of full CHARACTER SET and COLLATION support last month
inadvertently caused a regression in the handling of views. The intended
behavior (for now) is for Skeema to ignore views entirely, but due to a bug
Skeema would throw a fatal error upon encountering a view. This is now fixed.

Fixes issue #2. Thank you to @tomkrouper at @github for reporting!