-- Fix to support bit fields when generating DDL statements.
The entirety of the fix is to add a condition at line 158 of schema.go
to specify non-quoted handling for default bit values.

There is also an update to aTable() in test fixture generator to
add a bit column to the schema.

There is one test that is failing:
--- FAIL: TestTableAlterModifyColumn (0.00s)
	table_test.go:314: Incorrect number of table alters: expected 3, found 4

I think that the test is making assumptions about the number of columns in the table
fixture and I wasn't certain of exactly the right approach here. If you have an idea
I can take a stab at it, or feel free to modify the test :)