Table.CreateStatement: make behavior more reliable
Previously, this method could be called on a table that did not have a pre-
cached SHOW CREATE TABLE statement called, and it would instead return what
Tengo thinks the CREATE TABLE statement should be. In practice, this situation
does not occur, since tables get pre-cached values upon initial schema
introspection. In the event that the schema wasn't introspected, it is clearer
for a caller to explicitly call GeneratedCreateStatement instead. So
CreateStatement now panics if no pre-cached statement is available, to indicate
a bug is present in the calling code.