Major refactor to remove caching of instances and schemas
The caching scheme proved overly complex for insufficient benefit. The new
approach may result in slightly more queries, but the data is guaranteed to be
fresh. And to help compensate on performance, several Instance methods that
previously operated on Schema and Table structs now operate on names alone,
removing the need to have queried the full struct in some call paths.

This is a breaking API change, which is acceptable since Go La Tengo is still
pre-1.0. The following method signatures or return semantics changed:

* NewInstance() no longer de-duplicates instances, and no longer errors upon
  creating a duplicate instance with different auth or params.
* Instance.HasSchema() can now return an error if its query fails
* The following methods now take string args instead of structs:
  * Instance.ShowCreateTable()
  * Instance.TableSize()
  * Instance.TableHasRows()
  * Instance.DropSchema()
  * Instance.AlterSchema()
  * Instance.DropTablesInSchema()
  * Instance.CloneSchema()
* The following methods no longer possibly return an error:
  * NewSchemaDiff()
  * Schema.TablesByName()
  * Schema.Table()
* Schema.OverridesServerCharSet() now takes an Instance as an arg
* Schema.Tables() no longer exists, replaced with exported field Schema.Tables
* Schema.PurgeTableCache() no longer exists, not needed
* Schema.CachedCopy() no longer exists, not needed