pull, push, diff: move ignore-table handling to tengo.StatementModifiers
This commit updates tengo and moves handling of ignore-table regexes to
tengo.TableDiff.Statement(). This removes the need for type switches just
to obtain the table name in the command handlers.

This commit does not affect the lint or init command handlers, which still
need to handle ignore-table on their own, as they do not use SchemaDiff in
their behavior.