InnoDB tables: strip no-op clauses from SHOW CREATE TABLE
There are a few possible clauses in CREATE TABLE that have no effect whatsoever
for InnoDB tables, and are not reflected anywhere in information_schema, yet are
persisted to *.frm files and included in SHOW CREATE TABLE. Previously the
presence of these clauses would make Tengo think a table is using features that
are unsupported for diff'ing. They are now stripped entirely and ignored by
Tengo as if they were never present.

These clauses are very rare in practice, but there are two cases that sometimes
show up in real-world use:

* InnoDB tables that are compressed and then later uncompressed. This leaves
  dangling KEY_BLOCK_SIZE clauses on indexes. See MySQL bug tracker
  https://bugs.mysql.com/bug.php?id=70534 for more info.

* Some tools, such as phpMyAdmin, automatically insert USING BTREE clauses
  into indexes despite this clause having no effect.