Bug fix for NormalizeCreateOptions(): regexps must be ordered consistently
The regexp for catching KEY_BLOCK_SIZE on an index is fairly specific, to avoid
matching table-level KEY_BLOCK_SIZE. This regexp must be used *after* the one
handling USING BTREE, otherwise it won't match. This means a map is unsuitable,
since map iteration order is random. A slice of struct values is now used
instead.