diff/push: Preserve order of new secondary indexes
When adding new secondary indexes to an existing table, the generated ALTER
TABLE statement will now properly preserve the requested order in its ADD INDEX
clauses. Fixes #28.

The main fix and unit test occur at the library (Go La Tengo) level, and are in
https://github.com/skeema/tengo/commit/ba3e8ce6eba; meanwhile this skeema commit
simply updates that dep to bring in the fix, and also adds an integration test.