Further improve separation/ordering of AddForeignKey clauses
The previous logic ensured that a single alter would not contain both
DropForeignKey and AddForeignKey clauses. This helps with some problems, but
could still lead to issues if the "parent" side of a new foreign key is a new
table, or new columns, or a new index -- but not until later in the same
SchemaDiff.

This commit fixes that situation by always moving AddForeignKey clauses to a
separate TableDiff (one per table), and putting these at the end of the
SchemaDiff's list of TableDiffs.