diff/push: apply ADD FOREIGN KEY alters last, after other alters
This commit updates dep github.com/skeema/tengo in order to bring in a fix
for a foreign key operation edge case. ADD FOREIGN KEY operations now occur
in separate ALTERs (one per table) which are run after all other ALTERs in
the same schema. This way, if the new FKs rely on other changes in the same
diff -- such as referencing a new table, columns, or index -- they will still
succeed.

This commit adds a test for one specific case that previously would have
failed, even with foreign_key_checks=0: adding an FK that needs a new index
on the "parent" (referenced) table, where the parent table name is
alphabetically after the child table.

This does not yet address the situation of ordering constraints of foreign keys
that span multiple schemas, which is rare in practice and very difficult to
solve.