diff/push: unsafe statements now checked in pre-flight
Previously, if unsafe statements were not allowed by the config, they would
be skipped but any safe statements for the same schema would be allowed. This
can be problematic with foreign keys, though: we might skip an unsafe ALTER
that dropped a column but also added an index that is subsequently needed by
a foreign key created in a separate ALTER.

This situation should be exceedingly rare, but in any case, the correct
approach seems to be skipping all DDL in a schema if at least one forbidden
unsafe change is present. So with this commit, we now check for unsafe
statements prior to executing anything in push. Diff also behaves this way now
too; even though diff does not execute DDL regardless, it needs to be a "dry
run" of push that behaves in an equivalent manner.