Change the Flavor abstraction to include major and minor version numbers
Distributions/forks of MySQL are now considered Vendors, and a Flavor now
consists of a (Vendor, Major, Minor) tuple.

Flavor-specific logic is starting to be abstracted into methods on Flavor,
allowing more generic calls akin to "does this Flavor allow blobs to have
defaults?" rather than a hard-coded "is this Flavor MariaDB 10.2+?".

This is a breaking API change, which is acceptable since this package hasn't hit
its 1.0.0 release yet.