Move most flavor-specific logic to the display layer
All methods that generate DDL strings (including individual definitions or
clauses) now accept a Flavor arg, which can be used to change how the DDL is
formatted for a specific flavor. Previously, the logic was mostly handled upon
schema introspection, which made test fixtures awkward. The test fixture
functions now have "ForFlavor" variants that accept a flavor arg.

This is a breaking API change, which is acceptable since this package has not
yet hit the 1.0.0 milestone.