Avoid information_schema capitalization issues in MySQL 8.0+
MySQL 8.0 made a change regarding how information_schema queries work. From
https://dev.mysql.com/doc/refman/8.0/en/data-dictionary-usage-differences.html,

"Previously, result set headers of columns selected from INFORMATION_SCHEMA
tables used the capitalization specified in the query [...] As of MySQL 8.0,
these headers are capitalized [...] If necessary, a column alias can be used to
achieve a different lettercase."

This commit introduces column aliases (AS clauses) for all information_schema
queries that use sqlx's Select() to read into a struct, to avoid this problem.

This change is a prerequisite for MySQL 8.0 compatibility, but please note this
package does not yet support MySQL 8.0 in general; other changes will also be
necessary.