Schema introspection: A few more fixes towards MySQL 8.0 compat
Index order cannot be introspected from information_schema in 8.0, so a small
amount of SHOW CREATE TABLE parsing is now performed in flavors using the new
data dictionary.

Table-level collations of utf8mb4 are now also explicitly included in 8.0, even
if the default collation is used.

Not yet fixed in this commit: several new intricacies of column-level charset
and collations in 8.0.