Instance.DropTablesInSchema(): Retry deadlocks from mysql 8.0
With the new data dictionary in MySQL 8.0, deadlocks are frequently encountered
when attempting to concurrently drop two tables that have a foreign key
constraint between them. This commit gives DropTablesInSchema the ability to
serially retry any DROP TABLE queries that deadlocked, and adds a test case that
fails without this change.