pull, push, diff: Permit .skeema to override flavor
push/diff: If flavor is set in .skeema, use the value here instead of auto-
detecting from the target instance. This permits operators to manually configure
the flavor when needed, for example if running a custom fork/patch-set of the
mysql-server.

pull: Previously, if a flavor was specified in .skeema but could not be auto-
detected from the running instance, Skeema would remove the flavor from .skeema.
Now it leaves it alone, allowing operators to manually configure the flavor
without risk of `skeema pull` removing it.