init, pull: Always store default-character-set and default-collation
Previously, in schema-level .skeema files, init and pull only populated the
default-character-set and default-collation options if the schema-level values
differed from the server's global defaults. This is brittle, as the server
global default is not tracked by Skeema, and the default server global values
have changed in MySQL 8.0. As of this commit, the init command will always
populate the values for each schema; the pull command will now fill in missing
values for existing schemas.