Fix edge case bug in Instance.CanConnect()
If an instance had non-empty defaultParams, and a connection pool without a
default database had previously been established, CanConnect() would return
true even if the database was down. This commit fixes the bug so that
CanConnect() correctly returns false in this case.

This situation is infrequent in real-world use, but easier to hit in test
suites that intentionally shut down a DockerizedInstance temporarily.