Instance.CanConnect(): Prevent connection leak after previous fix
The previous commit's fix to Instance.CanConnect() causes a connection leak,
since each call potentially orphans the previous connection pool to create a
new one. The old pool would get cleaned up in 30 sec or less due to its max
connection lifetime, but given a lot of calls to CanConnect in a short time,
this can lead to hitting mysqld's max connections limit (depending what it is
configured to).

This commit uses a different approach to verify connectivity, which does not
require orphaning an existing connection pool.