Rename fs.IdealSchema and workspace.MaterializeIdealSchema
fs.IdealSchema is now fs.LogicalSchema, which better conveys the meaning of
the struct: it's a logical representation of the schema, in the same way that
a "logical backup" is a set of SQL that can be executed to recreate a physical
representation.

Along the same lines, workspace.MaterializeIdealSchema() is now called
workspace.ExecLogicalSchema().