Instance.querySchemaTables(): Fix naming conflict bug between FKs / UKs
Within a single database schema, if a foreign key has the same name as a unique
index, Instance.querySchemaTables() would consistently return an error. This
occurred because an additional join condition is needed to de-dupe this
situation, as both types of constraints are present in MySQL's
information_schema.key_column_usage table.

This commit does the following:

* Add the missing join condition

* Augment error messages in Instance.querySchemaTables() to include schema name

* Provide test coverage by renaming one of the FKs in the fixture data to match
  one of the unique keys