Fix error from foreign key with same name as unique key
Within a single database schema, if a foreign key has the same name as a unique
index, all Skeema commands would fail with this error message:

    Error querying foreign key constraints: sql: Scan error on column index 6:
    unsupported Scan, storing driver.Value type <nil> into type *string

This occurred because an additional join clause is needed to de-dupe this
situation, as both types of constraints are present in MySQL's
information_schema.key_column_usage table.

The bug and its fix are both actually in github.com/skeema/tengo, so this commit
just updates the vendored dep. Test coverage is handled by
https://github.com/skeema/tengo/commit/c9951fee4392c587aeccdcc90da2b184f5272f60#diff-ad27291d895bd1679307e9c94c251037

Fixes #46.