Add new method Instance.StrictModeCompliant()
This method returns true if all tables in the supplied schemas could be
recreated on the receiver instance even if innodb_strict_mode and a strict
sql_mode are in-use.

This commit also adds another supporting method, Table.RowFormatClause(),
as well as integration test coverage for recently-added method
Flavor.InnoRowFormatReqs().