DockerizedInstance: two small fixes for edge-cases
DockerizedInstance.SourceSQL() now specifically adds "-u root" to the mysql
client command-line. This is necessary in recent Percona releases, which
recently changed Docker user to "mysql" in PR
https://github.com/percona/percona-docker/pull/51. This means `docker exec`
also uses "mysql" user, which in turn acts as the default for mysql
connections too if no user is specified.

DockerizedInstance.Stop() increased timeout from 3sec to 10sec. In cases where
the timeout is hit, the instance may not be able to be started again, and this
seems to happen more frequently with 5.7 images (mysql or percona) on MacOS
for unknown reasons.