Internal improvement: Allow Workspace package to use third-party providers
This is an internal change that doesn't affect Skeema's functionality, but
makes one of its packages more flexible when used as an independent library.

Previously, workspace.ExecLogicalSchema() was only compatible with Workspace
types defined in the workspace package. This commit adds a new "pass-through"
type, workspace.TypePrefab, which allows use of any value satisfying the
workspace.Workspace interface.