push: Thread-safety fix for --concurrent-instances w/ --workspace=docker
The previous logic for --workspace=docker was not thread-safe, which could
cause problems if --concurrent-instances was set to 2 or more.

This commit also refactors shutdown handling of workspaces in general,
permitting more flexibility if the workspace package is used as a library
in another program. The previous shutdown model was only well-suited for
command-line applications / single processes, but not for a request-driven
architecture that needs to shut down workspaces used by only a single request.