diff/push: Fix handling of foreign key NO ACTION clauses
If a foreign key definition contained an "ON {UPDATE|DELETE} NO ACTION"
clause, Skeema would erroneously flag the table as using an unsupported
feature, preventing `skeema diff` and `skeema push` from operating on the
table. This bug affected all database flavors except for MySQL 8.

The actual fix (and corresponding test logic) is in
https://github.com/skeema/tengo/commit/9139dc8, so this commit just updates
the dep.

Fixes #50. Thank you to @mugli for the bug report!