Internal refactor: generic handling of different db object types
This commit refactors logic that formerly operated only on tables, so that
other object types can be supported with minimal changes in the future.

Additionally, this commit improves Skeema's ability to parse .sql files.
Although Skeema's logic largely does not rely on parsing SQL, minimal parsing
is still necessary to extract object names and types, and associate live db
objects with statements in files. However, the previous iteration of the
parser had several deficiencies, and was not easily extendable to support
additional types of objects. This commit replaces the parser with one based
on external dep github.com/alecthomas/participle (MIT license), which is now
vendored.

Finally, this commit introduces the ability to parse the DELIMITER command
in .sql files, which will be required to fully support multi-statement object
types in the future.

The changes in this commit were pulled out of the routines branch; this
essentially takes all non-routine-specific changes from that branch and merges
them early. This way, the routine-specific changes can be viewed independently,
serving as a simpler example for how to add other new object types to Skeema.