Add support for routines (stored procs and funcs)
This commit adds full support for stored procedures and functions to this
package. Routines are introspected by querying information_schema.routines,
as well as either mysql.proc (if available) or SHOW CREATE.

Routines can be diff'ed.  For now, any alter to a routine is handled by a DROP
and re-CREATE. (Most routine modifications require this anyway, but some
metadata-only changes can use ALTER FUNCTION or ALTER PROCEDURE in theory;
this may be implemented in the future.)

Full unit test and integration test coverage is included.

This commit is essentially a rebase of the routines branch onto master commit
5bedfdb68, which pulled out the non-routine-specific refactoring from that
branch.