pull: Don't leave file with only dangling commands
If a .sql file consists of 1 CREATE and 1 or more commands (USE or DELIMITER),
and the db object corresponding to the CREATE has been dropped out-of-band,
running `skeema pull` will now properly delete the .sql file. Previously, it
would leave the file in place with the CREATE removed. A .sql file consisting
only of commands serves no purpose, and creates clutter if the object name is
subsequently re-used.