diff/push: New option --compare-metadata for routine metadata
In MySQL and MariaDB, whenever a procedure, function, trigger, or event is
created, two pieces of environmental metadata are automatically stored and
associated with the object: the session sql_mode in effect at the time, and
the default collation of the database containing the object. This metadata
affects execution of the stored program in subtle ways.

This commit adds a new option, compare-metadata, controlling whether or not
Skeema should include this metadata in its diff logic. It is disabled by
default. (Note that previous pre-release builds of this branch effectively
had this behavior enabled, so this is a change in behavior.)

This commit also updates dep github.com/skeema/tengo to bring in the logic
supporting the new option.