Option file handling: ignore host option in .my.cnf
Skeema parses the standard per-user MySQL option file, ~/.my.cnf, for username
and password values. The intended behavior is to ignore the host option
though, if present in this specific file. However, a regression in Skeema
v1.0.6 broke this, and presence of host in ~/.my.cnf was triggering a fatal
error.

This commit fixes the regression, so that host is now ignored in ~/.my.cnf
again. See discussion in #64 for more context.