Instance and Flavor: more robustness around parsing problems
If the instance's vendor cannot be parsed from @@global.version_comment, try
parsing @@global.version as a fallback.

Add a new method Instance.SetFlavor(), which can override FlavorUnknown in
cases of parsing problems.

Add a new method Flavor.MySQLishMinVersion(), which does version comparisons
on non-MariaDB vendors. This method is now used in all checks that previously
looked explicitly for VendorMySQL and VendorPercona, meaning that these checks
will intentionally now be triggered if the vendor is unknown but the version
is sufficient.