Optimize information_schema query for foreign keys
Pre-mysql8, the query optimizer for some information_schema tables is quite
picky. To avoid scanning all databases, queries on key_column_usage must filter
on table_schema (NOT the completely-equivalent constraint_schema), and must
use a constant value (NOT derived from a join, even if evaluates to a
constant anyway).

This should improve performance on MySQL instances with a large number of
schemas (databases), regardless of whether or not they use foreign keys.