Internal improvements to directory management
Symlinks to .sql files now require that the *target* of the symlink be a
regular file with a .sql extension, rather than requiring that the *symlink*
itself has a .sql extension. The target must also be within the same
repository as the symlink, in the case of a Git repo. Otherwise, the target
must be within the same subdirectory structure based at either $HOME or the
root-most .skeema file found.

Linter configuration errors now include information on which directory had the
error, relative to the repo root.

A few other error messages that were previously relative to the current
working directory are now instead relative to the repo root.

This commit includes backwards-incompatible internal changes to the following
functions' signatures and/or logic:

* fs.Dir.RelPath()
* fs.ParentOptionFiles()
* linter.BadConfigResult()

Although this is typically avoided, there are no known third-party users of
these functions, so this change is considered safe.