Option files: sanity-check symlinks
This commit adds logic to ensure .skeema option files look sane, prior to
reading and parsing the file. Now .skeema must be a regular file or a symlink;
and if it is a symlink, it must point to a regular file, within the same
repo / dir tree.

Any subdirectory with a .skeema file not meeting these requirements is
ignored, with a warning logged. If the working directory's .skeema file does
not meet these requirements, it is fatal.

These changes are designed to improve safety of CI systems which integrate
Skeema in an automated fashion.

These restrictions do not apply to global option files, which are typically
outside of the repo anyway.