MySQL 8.0.13+: Support use of default expressions
MySQL/Percona 8.0.13 added support for default expressions; however, they are
reflected in information_schema.columns differently than MariaDB 10.2+'s
implementation. This commit adds support for both formats.

This also fixes support for columns with DEFAULT CURRENT_TIMESTAMP ON UPDATE
CURRENT_TIMESTAMP, which was incorrectly flagged as unsupported-for-diff in
MySQL/Percona 8.0.13+ with previous releases of this package.

This commit removes the method Flavor.AllowDefaultExpression(), since
determining availability of default expressions now requires looking at the
patch version as well, and the Flavor abstraction does not yet have a way of
capturing this.