Shellouts for host-wrapper and schema: ignore trailing spaces
The host-wrapper and schema options both support shelling out to an external
script, whose output is parsed to obtain a list of host:port or schema names,
respectively. When parsing the output, these options support a few different
delimiters (in order from highest precedence: newlines, commas, tabs, spaces)
and attempt to split the output on the highest-precedence delimiter found in
the output string.

However, since newline has the highest precedence, this means output with a
trailing newline would cause Skeema to assume newline is the delimiter, even
if the output was actually using some other delimiter.

This commit fixes the bug by removing trailing (and leading) spacing of any
type, prior to searching the output for delimiters.

Additionally, this commit fixes exposure of error messages around invalid
values for host and host-wrapper, which were inadvertently obscure ("dir maps
to an empty list of instances") previously in some cases.