push: delay instance schema introspection until just before diff
Previously, when `skeema push` would fully generate the list of (directory,
instance, schema) "target" tuples, it would introspect all schemas on the
instance at that time. This could result in a stale cache in a specific edge
case: if Skeema is operating on multiple targets, and one of the targets takes
a long time to push (e.g. altering a large table), while a subsequent target
is changed manually / out-of-band (e.g. someone ran DDL completely outside of
Skeema, while a `skeema push` process was running separately).

This commit delays the introspection of the current schema on each instance
until just before the push process needs to compute a diff, guaranteeing an
update-to-date view of the schema.

There are no known reports of the previous behavior being problematic, since
the edge case is rare and only occurs from questionable operational process.
Regardless, the newer approach is safer.