New command `skeema format` for reformatting files without linting
This commit adds a new command, `skeema format`, which reformats CREATE
statements in *.sql files to match the canonical format shown by SHOW CREATE.
Its behavior is similar to that of `skeema lint` in pre-1.2 releases, when the
linter did not yet search for table problems besides SQL errors.

This command exits with code 0 only if all statements are already using the
canonical format.

Users who want to check if statements are already formatted properly (but
without actually writing to files) may use the --skip-write option. This can
be useful in CI pipelines that verify proper formatting of commits, to enforce
a strict SQL style guide.