MySQL 8.0: fix ordering of table create options
In all previous versions of MySQL, the order of table option clauses in SHOW
CREATE TABLE exactly matched what was shown in
information_schema.tables.create_options. This is no longer the case in
MySQL 8.0 and Percona Server 8.0, due to the new internal data dictionary.

As a result, tables with multiple create options were sometimes erroneously
marked as unsupported-for-diff in these versions. This commit fixes the bug.