Add new --lint-auto-inc checker for auto_increment sanity
This option, along with corresponding whitelist --allow-auto-inc, provides a
mechanism for restricting what data types may be used for auto_increment
columns. This offers a way to avoid overly-small int types (which are prone
to overflow/max-out issues), signed int types (which have undefined auto_inc
behavior for negative numbers), and float types (which are nonsensical for
auto_inc and now deprecated in MySQL 8.0.17).

This option also provides a mechanism for banning use of auto_increment
altogether, by way of intentionally setting --allow-auto-inc to an empty list.