Routines: fix parse failure edge-case
Instance.Schemas() would fail with a parse error if any stored procs/funcs
contained newlines in their param list, AND the mysql.proc table was
unavailable (either due to privs, or due to not existing in MySQL 8.0+). The
root cause was a regular expression that did not account for how SHOW CREATE
PROCEDURE/FUNCTION retains user-supplied formatting of param lists.

The new solution avoids use of regular expressions for this code path
entirely.